fix: [ANDROAPP-3226] Filters allow unwanted event creation
